fix a bit

+++ backend/migrations/20220918060908_create_attachments.up.sql
@@ -1,7 +1,7 @@
 create table attachments (
-  id bigserial primary key,
+  id serial primary key,
   record_type varchar not null,
   record_id varchar not null,
   record_column_name varchar not null,
-  blob_id bigint references attachments(id) on delete cascade
+  blob_id int references attachments(id) on delete cascade
 );
